feat(register): get user inputs
Create a stateful component `[inputs, setInputs]` (a `state` variable)
that can handle changes to multiple input fields in a form by keeping
track of their values in state & updating that state whenever the user
types into one of the fields.

The `[inputs, setInputs]` state variable contains:
- `inputs` variable is the current state value, which is initially set
  to `{username: "", email: "", ...}`.
- `setInputs` variable is a function that can be used to update the
  state value.

By using state in this way, React can keep track of the values of the
input fields in the form, and re-render the component whenever the state
changes, ensuring that the component always reflects the current state
of the application.

Issue #64
Reference:
- [React.js State](https://www.simplilearn.com/tutorials/reactjs-tutorial/reactjs-state)