feat(register): persist registed user info to db
When `Register` button is clicked, client-side sends a POST request to
server-side with form data to persist registered info. Also, it updates
a state variable `err` with any error messages returned by the server.

`{err & err}` is a conditional rendering statement that checks if the
`err` state is truthy, i.e., not `null`, `undefined`, `false`, `0`,
`NaN`, or an empty string. If it is truthy, then the value of `err` is
rendered as a child of the form element.

Issue #64
Reference:
- [Await React](https://www.valentinog.com/blog/await-react/)